+++ src/aks-preview/azext_aks_preview/custom.py
@@ -604,6 +604,7 @@ def aks_create(
     enable_cilium_dataplane=False,
     custom_ca_trust_certificates=None,
     enable_advanced_network_observability=None,
+    advanced_networking_observability_tls_management=None,
     # nodepool
     crg_id=None,
     message_of_the_day=None,
@@ -836,6 +837,7 @@ def aks_update(
     safeguards_excluded_ns=None,
     enable_advanced_network_observability=None,
     disable_advanced_network_observability=None,
+    advanced_networking_observability_tls_management=None,
     # metrics profile
     enable_cost_analysis=False,
     disable_cost_analysis=False,